Allow using X-Accel for static user generated files
Motivator for this TIC is green IT by saving network traffic and CPU cycles.
The delivery of static files is a bottleneck for very few, but for us it
accounts for a significant portion of Stud.IP's network traffic (e.g.,
in the absence of a cloud solution for users and 20,000 students) and a
small but not overlooked portion of the CPU load of our PHP FPMs.

When a user clicks on a file to download, the request goes to the reverse
proxy, which forwards the request to a PHP FPM, which fetches the file from
a file server and returns the response to the reverse proxy, which returns
it to the user.

With X-accel this can be abbreviated: The proxy can deliver the file
directly after the authorization check by the PHP FPMs.

Resolves #2344